% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendi.R
\name{entropy_q}
\alias{entropy_q}
\title{entropy_q}
\usage{
entropy_q(K, q = 1)
}
\arguments{
\item{K}{A \eqn{n x n} normalized similarity matrix (must be Positive Semi-definite)}

\item{q}{Vendi score order, must be non-negative. \code{Default=1.}}
}
\value{
A single value describing the eigenvalues entropy of \code{K}
}
\description{
Returns the entropy of the eigenvalues of a given similarity matrix.
 This is equivalent to the log of the Vendi Score
}
