% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vendi.R
\name{weight_K}
\alias{weight_K}
\title{weight_K}
\usage{
weight_K(K, p = NULL)
}
\arguments{
\item{K}{A \eqn{n \times n} normalized similarity matrix (must be Positive Semi-definite)}

\item{p, }{an optional vector of length n must be non-negative and sum to 1. \code{Default=NULL}}
}
\value{
A weighted similarity matrix \code{K}
}
\description{
Computes a weighted similarity matrix.
}
